+++ example/config/settings.py
@@ -77,7 +77,6 @@ WSGI_APPLICATION = 'config.wsgi.application'
 DATABASES = {
     'default': {
         'ENGINE': 'clickhouse_backend.backend',
-        'NAME': 'test',
     }
 }