+++ tests/backends/clickhouse/test_creation.py
@@ -1,0 +1,89 @@
+from contextlib import contextmanager
+from io import StringIO
+from unittest import mock
+
+from clickhouse_driver.errors import ServerException, ErrorCodes
+from django.db import DatabaseError, connection
+from django.db.backends.base.creation import BaseDatabaseCreation
+from django.test import SimpleTestCase
+
+from clickhouse_backend.backend.creation import DatabaseCreation
+
+
+class DatabaseCreationTests(SimpleTestCase):
+    @contextmanager
+    def changed_test_settings(self, **kwargs):
+        settings = connection.settings_dict["TEST"]
+        saved_values = {}
+        for name in kwargs:
+            if name in settings:
+                saved_values[name] = settings[name]
+
+        for name, value in kwargs.items():
+            settings[name] = value
+        try:
+            yield
+        finally:
+            for name in kwargs:
+                if name in saved_values:
+                    settings[name] = saved_values[name]
+                else:
+                    del settings[name]
+
+    def _execute_raise_database_already_exists(self, cursor, parameters, keepdb=False):
+        server_error = ServerException(
+            "Database test already exists.",
+            ErrorCodes.DATABASE_ALREADY_EXISTS
+        )
+        error = DatabaseError(server_error)
+        raise DatabaseError(server_error) from error
+
+    def _execute_raise_unhandled(self, cursor, parameters, keepdb=False):
+        server_error = ServerException(
+            "",
+            ErrorCodes.DATABASE_ACCESS_DENIED
+        )
+        error = DatabaseError(server_error)
+        raise DatabaseError() from error
+
+    def patch_test_db_creation(self, execute_create_test_db):
+        return mock.patch.object(
+            BaseDatabaseCreation, "_execute_create_test_db", execute_create_test_db
+        )
+
+    @mock.patch("sys.stdout", new_callable=StringIO)
+    @mock.patch("sys.stderr", new_callable=StringIO)
+    def test_create_test_db(self, *mocked_objects):
+        creation = DatabaseCreation(connection)
+        # Simulate test database creation raising "database already exists"
+        with self.patch_test_db_creation(self._execute_raise_database_already_exists):
+            with mock.patch("builtins.input", return_value="no"):
+                with self.assertRaises(SystemExit):
+                    # SystemExit is raised if the user answers "no" to the
+                    # prompt asking if it's okay to delete the test database.
+                    creation._create_test_db(
+                        verbosity=0, autoclobber=False, keepdb=False
+                    )
+            # "Database already exists" error is ignored when keepdb is on
+            creation._create_test_db(verbosity=0, autoclobber=False, keepdb=True)
+        # Simulate test database creation raising unexpected error
+        with self.patch_test_db_creation(self._execute_raise_unhandled):
+            with mock.patch.object(
+                DatabaseCreation, "_database_exists", return_value=False
+            ):
+                with self.assertRaises(SystemExit):
+                    creation._create_test_db(
+                        verbosity=0, autoclobber=False, keepdb=False
+                    )
+                with self.assertRaises(SystemExit):
+                    creation._create_test_db(
+                        verbosity=0, autoclobber=False, keepdb=True
+                    )
+        # Simulate test database creation raising "insufficient privileges".
+        # An error shouldn't appear when keepdb is on and the database already
+        # exists.
+        with self.patch_test_db_creation(self._execute_raise_unhandled):
+            with mock.patch.object(
+                DatabaseCreation, "_database_exists", return_value=True
+            ):
+                creation._create_test_db(verbosity=0, autoclobber=False, keepdb=True)