+++ tests/backends/clickhouse/test_operations.py
@@ -1,0 +1,18 @@
+import unittest
+
+from django.core.management.color import no_style
+from django.db import connection
+from django.test import SimpleTestCase
+
+from ..models import Person, Tag
+
+
+class OperationsTests(SimpleTestCase):
+    def test_sql_flush(self):
+        self.assertEqual(
+            connection.ops.sql_flush(
+                no_style(),
+                [Person._meta.db_table, Tag._meta.db_table],
+            ),
+            ['TRUNCATE "backends_person"', 'TRUNCATE "backends_tag"'],
+        )