+++ tests/settings.py
@@ -1,11 +1,18 @@
 SECRET_KEY = 'fake-key'
-INSTALLED_APPS = [
-    'django.contrib.contenttypes',
-]
+INSTALLED_APPS = []
 DATABASES = {
     'default': {
         'ENGINE': 'clickhouse_backend.backend',
-        'NAME': 'test',
+        'NAME': 'default',
+        'OPTIONS': {
+            'settings': {
+                'mutations_sync': 1,
+            }
+        }
+    },
+    'other': {
+        'ENGINE': 'clickhouse_backend.backend',
+        'NAME': 'other',
         'OPTIONS': {
             'settings': {
                 'mutations_sync': 1,